Rename `Cell_map3_combo_tests` to `Cell_map3_state_tests`
